# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API",
        "description": "Documentação da API do projeto Weather",
        "contact": {
            "name": " - Lucas Vieira",
            "email": "lucasvieiraandrade58@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/registro": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registro de um novo usuário",
                "operationId": "efe613daaf2a04ed607bd06cfebeda71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "username"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadastro realizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsIn..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    },
                    "422": {
                        "description": "Requisição não pode ser processada."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticação de usuário",
                "operationId": "bf939b2261fe037ef9fcfc7c1f3e459a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsIn..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Encerra a sessão do usuário",
                "operationId": "846cd3967b46a23b4c10211e56743b21",
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/weather": {
            "get": {
                "tags": [
                    "Dados Meteorológicos"
                ],
                "summary": "Busca dados meteorológicos de uma cidade",
                "operationId": "43154892a0abd668626ce11c57fced9f",
                "parameters": [
                    {
                        "name": "cidade",
                        "in": "query",
                        "description": "Nome da cidade que deseja ver os dados meteorológicos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Requisição não pode ser processada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/weather/historico": {
            "get": {
                "tags": [
                    "Dados Meteorológicos"
                ],
                "summary": "Busca histórico de buscas meteorológicas de uma cidade",
                "operationId": "18f0e241865b39e72147ab6d05328a98",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "query",
                        "description": "Para filtrar por usuário",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "periodo",
                        "in": "query",
                        "description": "Para filtrar pela últimas 24 horas",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Requisição não pode ser processada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/weather/historico/{codigoHistorico}": {
            "get": {
                "tags": [
                    "Dados Meteorológicos"
                ],
                "summary": "Busca detalhes de um histórico",
                "operationId": "d4c55f16dccf63c084e68f831945cbfb",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Requisição não pode ser processada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "Dados Meteorológicos",
            "description": "Dados Meteorológicos"
        }
    ]
}